Allow multiple library name patterns for Logging.
This changeset makes it possible to specify a *WhiteList* of library
names for event logging. I wish to replace the current WhiteList
approach with a BlackList approach where we could just specify the
library patterns that we do _not_ wish to log/replay. The prominent
candidates for the blacklist would be libc, libpthread, libdl, libm,
dmtcphijack.so, fredhijack.so etc.

git-svn-id: svn://joust.ccs.neu.edu/home/tyler/svnroot/pthread-wrappers/fred@488 679428e0-fc1f-4e08-a6f4-36083810378a