Kill inferior first during restarts.
If we broadcast DMT_KILL_PEER message first, the inferior cannot handle the
signal because it is traced, but gdb does quit. Then the inferior is stuck
indefinitely in the traced state. The solution is to send SIGKILL to the
inferior first, so that the inferior does not have to handle the DMT_KILL_PEER
message at all.