Fix race cond. capturing child output: for r-w.
There was a race that could affect reverse-watch. If the "evaluate_expression"
function was called while the child (debugger) was still printing output to the
screen, the saved value of the expression at the start of reverse-watch could
contain extraneous output from the child, or it could fail to capture the
expression value at all. This could cause reverse-watch to be unable to
determine an interval in which the expression changes value.